Refactor: move request model to request package and rename transformer
